memoristaIA-main/requirements.txt
Django
djangorestframework
google-generativeai
python-dotenv

